\name{BiocStyle-deprecated}

\alias{BiocStyle-deprecated}
\alias{latex_old}
\alias{latex2}
\alias{pdf_document_old}
\alias{pdf_document2}
\alias{html_document_old}
\alias{html_document2}

\title{Deprecated functions in package \sQuote{BiocStyle}}

\description{
  These functions are provided for compatibility with older versions
  of \sQuote{BiocStyle} only, and will be defunct at the next release.
}

\details{
  The following functions are deprecated and will be made defunct; use
  the replacement indicated below:
  \itemize{
    \item{latex_old, latex2: \code{\link{latex}}}
    \item{pdf_document_old, pdf_document2: \code{\link{pdf_document}}}
    \item{html_document_old, html_document2: \code{\link{html_document}}}
  }
}

\usage{
latex_old(..., width=90, short.fignames=FALSE, fig.path, error=FALSE,
          use.unsrturl=TRUE)

pdf_document_old(toc = TRUE, number_sections = TRUE, use.unsrturl = TRUE, 
                 includes, ...)
                 
html_document_old(toc = TRUE, toc_depth = 3, number_sections = TRUE, 
                  fig_width = 7, fig_height = 5, fig_retina = NULL,
                  fig_caption = FALSE, dev = "png", smart = TRUE, 
                  self_contained = TRUE,  highlight = "default", 
                  mathjax = "default", extra_dependencies = NULL, css = NULL,
                  includes = NULL, keep_md = FALSE, lib_dir = NULL,
                  md_extensions = NULL, pandoc_args = NULL, ...)
}

\arguments{

  \item{\dots}{Additional arguments, passed to other methods}

  \item{width}{integer(1), maximum number of columns on a line used in
    printing. See \code{\link{options}}.}

  \item{short.fignames}{logical(1), indicates whether \code{\\incfig}
    figures should be inserted and referred to using short names equal to
    corresponding code chunk labels without any additional prefix.}

  \item{fig.path}{character(1), custom prefix to be used for figure file
    names when used with \code{knitr}; has no effect when compiled with
    \code{Sweave}. For details see \code{\link[knitr]{opts_chunk}}.}

  \item{error}{logical(1), behavior of \code{knitr} when errors occur in
    vignette code chunks, has no effect when compiled with
    \code{Sweave}. The default (\code{FALSE}) means that errors stop
    vignette processing, which is appropriate when packages are being
    built and checked. For details see \code{\link[knitr]{opts_chunk}}.}
    
  \item{use.unsrturl}{logical(1), indicating that the \sQuote{unsrturl}
    style will be used (\code{\\bibliographystyle} command \emph{not} required).}

  \item{toc}{logical(1), \code{TRUE} to include a table of contents in the output.}

  \item{number_sections}{logical(1), \code{TRUE} to number section headings.}
    
  \item{includes}{Named list of additional content to include within the document}

  \item{toc_depth}{Depth of headers to include in table of contents}
  
  \item{fig_width}{Default width (in inches) for figures}
  
  \item{fig_height}{Default width (in inches) for figures}
  
  \item{fig_retina}{Scaling to perform for retina displays}
  
  \item{fig_caption}{\code{TRUE} to render figures with captions}
  
  \item{dev}{Graphics device to use for figure output (defaults to png)}
  
  \item{smart}{Produce typographically correct output, converting straight
  quotes to curly quotes, --- to em-dashes, -- to en-dashes, and ... to
  ellipses.}
  
  \item{self_contained}{Produce a standalone HTML file with no external
  dependencies, using data: URIs to incorporate the contents of linked
  scripts, stylesheets, images, and videos. Note that even for self
  contained documents MathJax is still loaded externally (this is
  necessary because of it's size).}
  
  \item{highlight}{Syntax highlighting style. Supported styles include
  "default", "tango", "pygments", "kate", "monochrome", "espresso",
  "zenburn", "haddock", and "textmate". Pass \code{NULL} to prevent syntax
  highlighting.}
  
  \item{mathjax}{Include mathjax. The "default" option uses an https URL from
  the official MathJax CDN. The "local" option uses a local version of
  MathJax (which is copied into the output directory). You can pass an
  alternate URL or pass \code{NULL} to exclude MathJax entirely.}
  
  \item{css}{One or more css files to include}
  
  \item{keep_md}{Keep the markdown file generated by knitting.}
  
  \item{lib_dir}{Directory to copy dependent HTML libraries (e.g. jquery,
  bootstrap, etc.) into. By default this will be the name of the document
  with \code{_files} appended to it.}
  
  \item{md_extensions}{Markdown extensions to be added or removed from the
  default definition or R Markdown. See the \code{\link[rmarkdown]{rmarkdown_format}}
  for additional details.}
  
  \item{pandoc_args}{Additional command line options to pass to pandoc}
  
  \item{extra_dependencies}{Additional function arguments to pass to the
  base R Markdown HTML output formatter \code{\link[rmarkdown]{html_document_base}}}
}
