% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_vignette.R
\name{use_vignette_html}
\alias{use_vignette_html}
\alias{use_vignette_pdf}
\title{Create 'Rmd' vignette templates for HTML or PDF output}
\usage{
use_vignette_html(destination = tempfile(fileext = ".Rmd"))

use_vignette_pdf(destination = tempfile(fileext = ".Rmd"))
}
\arguments{
\item{destination}{character(1) file path to destination. The 
directory of the destination must already exist, and the file 
must not. The default creates a file in the temporary 
directory, and so is removed when the R session ends.}
}
\description{
Create 'Rmd' vignette templates for HTML or PDF output
}
\examples{
use_vignette_html()

}
