\name{BiocStyle-deprecated}

\alias{BiocStyle-deprecated}
\alias{latex_old}
\alias{latex2}
\alias{pdf_document_old}
\alias{pdf_document2}
\alias{html_document_old}
\alias{html_document2}

\title{Defunct functions in package \sQuote{BiocStyle}}

\description{These functions are defunct and no longer available.}

\details{
  The following functions are no longer available; use
  the replacement indicated below:
  \itemize{
    \item{latex_old, latex2: \code{\link{latex}}}
    \item{pdf_document_old, pdf_document2: \code{\link{pdf_document}}}
    \item{html_document_old, html_document2: \code{\link{html_document}}}
  }
}
