% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiocStyle-defunct.R
\name{BiocStyle-defunct}
\alias{BiocStyle-defunct}
\title{Defunct functions in package \sQuote{BiocStyle}}
\description{
These functions are defunct and no longer available.
}
\details{
The following functions are no longer available; use the replacement
indicated below:
\itemize{
  \item \code{latex_old, latex2: \link{latex}}
  \item \code{pdf_document_old, pdf_document2: \link{pdf_document}}
  \item \code{html_document_old, html_document2: \link{html_document}}
}
}
