% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\name{helpers}
\alias{helpers}
\alias{doc_date}
\alias{pkg_ver}
\title{Specify Rmarkdown document metadata}
\arguments{
\item{pkg}{character(1), package name}
}
\value{
Markdown-formatted character string.
}
\description{
Helper functions for including metadata in the document header.
}
\details{
Use \code{doc_date} to include document compilation date in the document
metadata field 'date', and \code{pkg_ver} for package version specification
in the 'package' field.
}
\examples{


## current date
doc_date()

## package name with version
pkg_ver("BiocStyle")

}
\author{
Andrzej Oleś <andrzej.oles@embl.de>, 2014-2015
}
\keyword{manip}
