\name{html_document}

\alias{html_document}

\title{Use Bioconductor style to format R Markdown HTML output}

\description{
Format for converting from R Markdown to an Bioconductor HTML document.
}

\usage{
html_document(toc = TRUE, toc_depth = 3, number_sections = TRUE,
  fig_width = 7, fig_height = 5, fig_retina = NULL,
  fig_caption = FALSE, dev = "png", smart = TRUE, self_contained = TRUE,
  highlight = "default", mathjax = "default",
  css = NULL, includes = NULL, keep_md = FALSE,
  lib_dir = NULL, md_extensions = NULL, pandoc_args = NULL, ...)
}
\arguments{
\item{toc}{\code{TRUE} to include a table of contents in the output}

\item{toc_depth}{Depth of headers to include in table of contents}

\item{number_sections}{\code{TRUE} to number section headings}

\item{fig_width}{Default width (in inches) for figures}

\item{fig_height}{Default width (in inches) for figures}

\item{fig_retina}{Scaling to perform for retina displays}

\item{fig_caption}{\code{TRUE} to render figures with captions}

\item{dev}{Graphics device to use for figure output (defaults to png)}

\item{smart}{Produce typographically correct output, converting straight
quotes to curly quotes, --- to em-dashes, -- to en-dashes, and ... to
ellipses.}

\item{self_contained}{Produce a standalone HTML file with no external
dependencies, using data: URIs to incorporate the contents of linked
scripts, stylesheets, images, and videos. Note that even for self
contained documents MathJax is still loaded externally (this is
necessary because of it's size).}

\item{highlight}{Syntax highlighting style. Supported styles include
"default", "tango", "pygments", "kate", "monochrome", "espresso",
"zenburn", "haddock", and "textmate". Pass \code{NULL} to prevent syntax
highlighting.}

\item{mathjax}{Include mathjax. The "default" option uses an https URL from
the official MathJax CDN. The "local" option uses a local version of
MathJax (which is copied into the output directory). You can pass an
alternate URL or pass \code{NULL} to exclude MathJax entirely.}

\item{css}{One or more css files to include}

\item{includes}{Named list of additional content to include within the
document (typically created using the \code{\link[rmarkdown]{includes}} function).}

\item{keep_md}{Keep the markdown file generated by knitting.}

\item{lib_dir}{Directory to copy dependent HTML libraries (e.g. jquery,
bootstrap, etc.) into. By default this will be the name of the document
with \code{_files} appended to it.}

\item{md_extensions}{Markdown extensions to be added or removed from the
default definition or R Markdown. See the \code{\link[rmarkdown]{rmarkdown_format}}
for additional details.}

\item{pandoc_args}{Additional command line options to pass to pandoc}

\item{...}{Additional function arguments to pass to the base R Markdown HTML
  output formatter}
}

\value{
R Markdown output format to pass to \code{\link[rmarkdown]{render}}
}

\details{
The \code{BiocStyle::html_document} format is based on the \code{\link[rmarkdown]{html_document}} format.
See the \href{http://rmarkdown.rstudio.com/html_document_format.html}{online
documentation} for additional details on using the \code{html_document} format.

R Markdown documents can have optional metadata that is used to generate a
document header that includes the title, author, date, package info and abstract. For more details
see the documentation on R Markdown \code{\link[rmarkdown]{metadata}}.

R Markdown documents also support citations. You can find more information on
the markdown syntax for citations in the
\href{http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html}{Bibliographies
and Citations} article in the online documentation.
}


\author{
  Andrzej Oleś <andrzej.oles@embl.de>, 2014-2015
}

\seealso{
  \code{\link[BiocStyle]{pdf_document}}, \code{\link[BiocStyle]{md_document}}
}

\examples{
\dontrun{

# simple invocation
render("input.Rmd", BiocStyle::html_document())

# specify options
render("input.Rmd", BiocStyle::html_document(toc = FALSE))
}
}

