% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\name{output}
\alias{output}
\title{Output format of an R Markdown document}
\usage{
output()
}
\value{
A character string specifying the pandoc output format.
}
\description{
Helper function to determine the document's current pandoc output format.
}
\details{
The function is useful for defining different behavior depending on the
output format, e.g. figure settings.
}
\examples{

\dontrun{
## Switch between SVG and PDF figures depending on document output format
knitr::opts_chunk$set(
  dev = switch(output(), html = "svg", latex = "pdf")
  )
}

}
\author{
Andrzej Oleś <andrzej.oles@embl.de>, 2016
}
