% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_vignette.R
\name{use_vignette}
\alias{use_vignette}
\alias{use_vignette_rmd}
\alias{use_vignette_pdf}
\title{Create vignettes that utilize the BiocStyle templates}
\usage{
use_vignette_rmd(vignette_name)

use_vignette_pdf(vignette_name)
}
\arguments{
\item{vignette_name}{character, the name of the vignette}
}
\value{
logical, indicating the vignette file was created successfully
}
\description{
Functions to create a skeleton Rmd or pdf vignette file.
}
\examples{
use_vignette_rmd("testpkg_vignette")

use_vignette_pdf("testpkg_vignette")
}
