% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_vignette.R
\name{use_vignette}
\alias{use_vignette}
\alias{use_vignette_html}
\alias{use_vignette_pdf}
\alias{use_package_vignette_html}
\alias{use_package_vignette_pdf}
\title{Create 'Rmd' vignette templates for HTML or PDF output}
\usage{
use_vignette_html(destination = tempfile(fileext = ".Rmd"))

use_vignette_pdf(destination = tempfile(fileext = ".Rmd"))

use_package_vignette_html(
  destination = tempfile(fileext = ".Rmd"),
  package = "MyPackage",
  title = paste("Introduction to", package),
  abstract = "Describe your vignette with a paragraph-length summary.",
  authors = person(),
  bug_report_url = "https://support.bioconductor.org"
)

use_package_vignette_pdf(
  destination = tempfile(fileext = ".Rmd"),
  package = "MyPackage",
  title = paste("Introduction to", package),
  abstract = "Describe your vignette with a paragraph-length summary.",
  authors = person(),
  bug_report_url = "https://support.bioconductor.org"
)
}
\arguments{
\item{destination}{character(1) file path to destination, usually
inside the `"vignettes"` directory of `package`. The directory
of the destination must already exist, and the file must
not. The default creates a file in the temporary directory, and
so is removed when the R session ends.}

\item{package}{character(1) package name.}

\item{title}{character(1) vignette title.}

\item{abstract}{character(1) paragraph-length summary of the vignette.}

\item{authors}{person() vignette authors. Use `person(comment =
...)` to include author affiliation (see examples).}

\item{bug_report_url}{character(1) url for bug reports (e.g.,
`"https://github.com/Bioconductor/BiocStyle/issues"`)}
}
\description{
`use_vignette_html()` and `use_vignette_pdf()` create
    markdown vignettte templates for general use, in html or pdf
    formats. The template outlines markdown syntax and unique
    features enabled by BiocStyle.

`use_package_vignette_html()` and
    `use_package_vignette_pdf()` create vignette templates suitable
    for inclusion in Bioconductor packages. The templates contain
    package installation instructions, suggest overall structure,
    and provide locations to seek support and create bug reports.
}
\examples{
vignette <- use_package_vignette_html(
    authors = c(
        person(
            "Iman", "Author",
            email = "iman.author@some.where",
            comment = c(affiliation = "Prestigious Institution")
        ),
        person("Iman", "Other")
    )
)
vignette
cat(readLines(vignette), "\n")
}
